Move library and update for conventions
