+++ source/bin_shell/CMakeLists.txt
@@ -17,6 +17,7 @@ if(WIN32)
         potato::spud
         potato::librender
         potato::libecs
+        potato::reflex
         glm
         SDL2
         SDL2main
