+++ source/reflex/public/potato/reflex/_wrapper.h
@@ -2,8 +2,8 @@
 
 #pragma once
 
-#include <potato/foundation/zstring_view.h>
-#include <potato/foundation/traits.h>
+#include <potato/spud/zstring_view.h>
+#include <potato/spud/traits.h>
 
 namespace up::reflex::_detail {
     // Creates the "public" interface inside of a reflect::serialize_value function
