+++ source/reflex/public/potato/reflex/reflect.h
@@ -7,9 +7,9 @@
 #include "_tag.h"
 #include "metadata.h"
 #include "traits.h"
-#include <potato/foundation/zstring_view.h>
-#include <potato/foundation/string.h>
-#include <potato/foundation/traits.h>
+#include <potato/spud/zstring_view.h>
+#include <potato/spud/string.h>
+#include <potato/spud/traits.h>
 
 namespace up::reflex {
